+++ flask_security_bundle/models/__init__.py
@@ -1,3 +1,3 @@
-from .user import User
+from .user import AnonymousUser, User
 from .role import Role
 from .user_role import UserRole