+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * use the default behavior of controllers to name endpoints (in effect, rename the prefixes from `security` to `security_controller`)
 * rename a translations keys for template h1 headings
 * move setting of the `csrf_token` cookie to flask unchained's controller bundle
+* do not depend on `UserMixin`, `RoleMixin`, or `AnonymousUser` from flask_security
+* do not depend on `login_user`, `logout_user`, `hash_password`, `get_hmac`, and `user_double_hash` utility functions from flask_security
 
 ## 0.3.0 (2018/07/25)